Use black color for X-axis labels (same as Y-axis)
The X-axis labels are in the "windowText" color, which depends on the
color scheme. The background color, however is always white. In
inverted color schemes, this text can become difficult to read, or
invisible (in particular, in the Breeze Dark High Contrast scheme
shipped with KDE).

The color is chosen to be black, the same choice as made with the
Y-labels.