Tag 0.9.1
Change of maintainer.  Thank you to M G Berberich for writing this
software!  A testament to the quality of the code is how little
maintenance it has needed in the last 9 years. However, the code
has several bugs and should be ported to newer QT.

This patch rolls changes that have been included in Debian, or in
my own version, into a new release.